# docs/openapi/ispwatch-partner-v1.yaml
...
openapi: 3.0.3
info:
  title: ISPWatch Partner API
  # 1.1.0 (2026-09-30): tipos de evento CUSTOMER_DELETED / ROUTER_CHANGED /
  # NETWORK_CHANGED, paginación por cursor `after_id`, filtro `router_id` en
  # /services y publicación en serie del feed. Todo aditivo: nada que ya se
  # consumía cambió de forma ni de significado.
  version: "1.1.0"
  description: |
    API de **solo lectura** para integradores externos (CRM, facturación
    electrónica, orquestadores AAA). Expone la autoridad comercial de ISPWatch
    —clientes, servicios contratados, cartera, soporte— y un feed de cambios
    para sincronizar sin re-consultar todo.

    ## Autenticación

    Llave de API en el header, emitida por el ISP o por el operador:

    ```
    Authorization: Bearer 42|xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
    ```

    La llave se muestra **una sola vez** al emitirla. Cada llave lleva:

    - un **tenant** fijo: sólo ve los datos de ese ISP, y no hay parámetro que
      lo cambie;
    - una **allowlist de IP** obligatoria: desde otra IP responde `403`
      `ip_not_allowed`;
    - un conjunto de **permisos** (abilities) por área;
    - un **vencimiento** (obligatorio si el ISP la emitió él mismo, máximo 90
      días).

    ## Reglas que no se negocian

    | Regla | Consecuencia |
    |---|---|
    | Sólo GET | Cualquier otro verbo responde `405 method_not_allowed`, aunque la llave sea válida |
...
    | Un tenant por llave | Un recurso de otro tenant responde `404`, nunca `403` |

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /customers:
    get:
      tags: [Clientes]
      operationId: listCustomers
      summary: Listar clientes
      description: |
        Abonados del tenant, ordenados por id ascendente.

        El filtro `document` es de **coincidencia exacta**, nunca parcial: una
        búsqueda por prefijo convertiría esta ruta en un enumerador de la base
        de clientes del ISP.
      x-ability: read:customers
      security:
        - bearerAuth: [read:customers]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - $ref: "#/components/parameters/AfterId"
        - $ref: "#/components/parameters/UpdatedSince"
        - name: service_status
          in: query
          description: Estado comercial exacto.
          schema: { $ref: "#/components/schemas/ServiceStatus" }
        - name: router_id
          in: query
          description: Router/grupo lógico al que pertenece el cliente.
          schema: { type: integer }
        - name: document
          in: query
          description: Documento/cédula. Coincidencia exacta.
          schema: { type: string, maxLength: 40 }
      responses:
        "200":
          description: |
            Página de clientes. `meta` es `PageMeta` si se pidió por `page`, y
            `KeysetMeta` si se pidió por `after_id`.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Customer" }
                  meta:
                    oneOf:
                      - { $ref: "#/components/schemas/PageMeta" }
                      - { $ref: "#/components/schemas/KeysetMeta" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
...
  /services:
    get:
      tags: [Servicios]
      operationId: listServices
      summary: Listar servicios contratados
      description: |
        El `service_id` de esta API es `user_services.id`: identificador
        estable y distinto del cliente.

        **Advertencia de mapeo.** Existe otra columna llamada
        `customer_profile.service_id` que **no es un servicio: es el plan**.
        Mapearla como identificador de servicio rompe la integración en
        silencio, sin ningún error visible.

        Hoy un cliente tiene un solo punto de red: los atributos de red
        (`ip`, `pppoe_username`, `router_id`) viven en la ficha del cliente,
        no en el servicio. El identificador propio existe para que el día que
        haya multipunto el contrato no tenga que cambiar.
      x-ability: read:services
      security:
        - bearerAuth: [read:services]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - $ref: "#/components/parameters/AfterId"
        - $ref: "#/components/parameters/UpdatedSince"
        - $ref: "#/components/parameters/CustomerId"
        - name: status
          in: query
          description: Estado del contrato de servicio.
          schema: { $ref: "#/components/schemas/ServiceContractStatus" }
        - name: service_status
          in: query
          description: Estado comercial del cliente dueño del servicio.
          schema: { $ref: "#/components/schemas/ServiceStatus" }
        - name: router_id
          in: query
          description: |
            Router/grupo lógico del servicio (`network.router_id`). Para
            sincronizar un NAS sin bajarse todos los servicios del ISP.
          schema: { type: integer }
      responses:
        "200":
          description: |
            Página de servicios. `meta` es `PageMeta` si se pidió por `page`,
            y `KeysetMeta` si se pidió por `after_id`.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Service" }
                  meta:
                    oneOf:
                      - { $ref: "#/components/schemas/PageMeta" }
                      - { $ref: "#/components/schemas/KeysetMeta" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Llave de API de ISPWatch. Formato `<id>|<secreto>`; se envía completa.
        Los permisos listados en cada operación son los `abilities` que la
        llave debe tener; sin ellos la respuesta es `403`.
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PerPage:
      name: per_page
      in: query
      description: Tope duro de 100.
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
    AfterId:
      name: after_id
      in: query
      description: |
        Paginación por cursor: devuelve las filas con id estrictamente mayor,
        en orden ascendente. `0` empieza desde el principio. La respuesta trae
        `KeysetMeta`; se sigue con `after_id = meta.next_after_id` mientras
        `meta.has_more` sea `true`.

        **Es el modo recomendado para recorrer una colección completa.** Con
        `page`, si se elimina una fila anterior a la página actual durante el
        recorrido, la paginación se corre y una fila que no cambió se salta
        sin ningún error. No se combina con `page` (`422`).
      schema: { type: integer, minimum: 0 }
    UpdatedSince:
      name: updated_since
      in: query
      description: |
        Devuelve sólo lo modificado desde esta fecha/hora. Sirve para
        sincronización incremental barata, pero **no reemplaza a `/events`**:
        dos escrituras en el mismo segundo son indistinguibles por reloj.

        **No sirve para decisiones de acceso.** En `/customers` compara la
        fecha del usuario, que no se mueve al cortar o reactivar al cliente;
        en `/services`, la del contrato de servicio, que no se mueve cuando
        cambia `service_status`.
      schema: { type: string, format: date-time }
    CustomerId:
      name: customer_id
      in: query
      schema: { type: integer }
...
  responses:
    Unauthorized:
      description: |
        Llave ausente, inexistente, vencida, revocada, o autenticada por un
        medio que no es una llave de API (por ejemplo una sesión del panel).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            revocada:
              value: { error: key_revoked, message: "Esta llave de API fue revocada." }
            vencida:
              value: { error: key_expired, message: "Esta llave de API está vencida." }
    Forbidden:
      description: |
        La llave es válida pero la petición no procede. El campo `error`
        distingue el motivo sin necesidad de llamar por teléfono.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            ip:
              value: { error: ip_not_allowed, message: "La IP de origen no está autorizada para esta llave." }
            permiso:
              summary: Falta el ability del área (respuesta de Sanctum)
              value: { message: "Invalid ability provided." }
            desactivado:
              value: { error: client_disabled, message: "El cliente de API está desactivado." }
            https:
              value: { error: https_required, message: "La API sólo se atiende sobre HTTPS." }
...
    ValidationError:
      description: Parámetros inválidos.
      content:
        application/json:
          schema:
            type: object
            properties:
              message: { type: string }
              errors:
                type: object
                additionalProperties:
                  type: array
                  items: { type: string }
    TooManyRequests:
      description: Se superó 60/min o 5.000/hora para esta llave.
      headers:
        Retry-After:
          description: Segundos a esperar antes de reintentar.
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    ServiceStatus:
      type: string
      description: |
        **Autoridad comercial de conectividad.** Es el campo que decide si hay
        servicio, y el que un integrador debe mirar.

        - `activo`, `gratis`, `suspendido` → siguen facturando;
        - `retirado`, `cancelado` → bajas definitivas.

        `suspendido` significa cortado por mora y **sigue generando factura**:
        no es una baja.
      enum: [activo, gratis, suspendido, retirado, cancelado]
    ServiceContractStatus:
      type: string
      description: |
        Estado del contrato de servicio, distinto del estado comercial del
        cliente. `gratis` es el servicio de cortesía.
      enum: [active, gratis]
...
    Decimal:
      type: string
      description: |
        Decimal en texto, con dos posiciones (`"85000.00"`). Se transporta como
        cadena a propósito: convertirlo a float binario introduce error de
        redondeo en cartera. Parsear con tipo decimal.
      example: "85000.00"
    Revision:
      type: integer
      nullable: true
      description: |
        `event_id` del último evento publicado de este recurso. Sirve para
        reconciliar sin depender del reloj: dos escrituras en el mismo segundo
        son indistinguibles por `updated_at`, por `revision` no.

        En `/services` es la revisión del **cliente** dueño: los atributos de
        red viven en su ficha, así que un cambio de router, IP o estado del
        cliente es un cambio de su servicio.

        `null` = el recurso no ha cambiado desde que existe el feed. **No es un
        error** ni un dato faltante.
    Error:
      type: object
      properties:
        error:
          type: string
          description: |
            Código estable y legible por máquina. Los rechazos que produce
            Sanctum (falta de ability) traen sólo `message`.
          enum:
            - method_not_allowed
            - https_required
            - invalid_credentials
            - client_disabled
            - tenant_missing
            - key_revoked
            - key_expired
            - ip_not_allowed
            - not_found
        message:
          type: string
          description: Explicación legible por una persona.
    PageMeta:
      type: object
      properties:
        page: { type: integer, example: 1 }
        per_page: { type: integer, example: 50 }
        total: { type: integer, example: 137 }
        last_page: { type: integer, example: 3 }
    KeysetMeta:
      type: object
      description: Paginación por cursor (`after_id`).
      properties:
        per_page: { type: integer, example: 100 }
        after_id: { type: integer, description: El cursor recibido., example: 0 }
        next_after_id:
          type: integer
          description: |
            Cursor para la página siguiente. Si la página vino vacía, devuelve
            el mismo `after_id` recibido.
          example: 1843
        has_more:
          type: boolean
          description: "`true` = hay más filas después de esta página."
...
    Plan:
      type: object
      nullable: true
      description: Plan comercial contratado. `null` si el cliente no tiene plan asignado.
      properties:
        id: { type: integer }
        name: { type: string, example: "Internet Fibra 10MB" }
        speed_down:
          type: string
          description: |
            **Texto, no número.** Se guarda como lo escribe el ISP (`"10M"`,
            `"10240k"`), porque es lo que se aplica literalmente en el equipo.
          example: "10M"
        speed_up: { type: string, example: "10M" }
        price:
          type: integer
          description: Precio mensual del plan, en la moneda del ISP.
          example: 142800
        is_courtesy:
          type: boolean
          description: |
            Plan de cortesía. Sólo presente en `/services`. Un cliente en un
            plan de cortesía no genera cobro aunque su estado sea `activo`.
    Customer:
      type: object
      properties:
        id: { type: integer, description: "Id del cliente (`users.id`)." }
        name: { type: string }
        last_name: { type: string, nullable: true }
        is_company: { type: boolean }
        document: { type: string, nullable: true, description: Cédula o NIT. }
        email:
          type: string
          nullable: true
          description: Correo de contacto del abonado, no su usuario de acceso al portal.
        phone: { type: string, nullable: true }
        address: { type: string, nullable: true }
        city: { type: string, nullable: true }
        state: { type: string, nullable: true }
        country: { type: string, nullable: true }
        postal_code: { type: string, nullable: true }
        latitude: { type: string, nullable: true, example: "4.45859700" }
        longitude: { type: string, nullable: true, example: "-74.63904800" }
        installation_date: { type: string, format: date, nullable: true, example: "2026-08-18" }
        is_enabled:
          type: boolean
          description: |
            Indicador heredado de "cliente habilitado". La autoridad es
            `service_status`, pero hoy los dos se mueven juntos: el corte (por
            mora o manual) lo pone en `false` y la reconexión en `true`.

            Una integración fail-closed debe exigir **ambos**
            (`service_status` en `activo`/`gratis` **y** `is_enabled = true`):
            puede haber datos anteriores a `service_status` donde no
            coinciden. Si cambia sin que cambie `service_status`, el feed
            emite `CUSTOMER_UPDATED` con `is_enabled` en `fields`.
        service_status: { $ref: "#/components/schemas/ServiceStatus" }
        credit_balance:
          allOf: [{ $ref: "#/components/schemas/Decimal" }]
          description: Saldo a favor del cliente.
        excluded_from_billing:
          type: boolean
          description: |
            `true` = el cliente está **fuera del ciclo automático completo**:
            no se le emite factura, ni recordatorio, ni aviso, ni se le corta.
            Precede a todo lo demás.
        is_fiber: { type: boolean }
        ip: { type: string, nullable: true, description: IP asignada en el router. }
        pppoe_username: { type: string, nullable: true }
        plan: { $ref: "#/components/schemas/Plan" }
        notify_invoice:
          type: boolean
          description: |
            Si ISPWatch le avisa la factura a este cliente.

            **Es por cliente, no por router.** Lo que se configura a nivel de
            router es el *canal* (email, WhatsApp o ambos), no el *si*. Hoy
            **no existe** una política de «no avisar factura» a nivel de
            grupo/router. Un integrador que emita el documento electrónico por
            su cuenta necesita este campo para no duplicarle el aviso al
            abonado.
        sectorial: { type: string, nullable: true, description: Sector/antena o NAP de fibra. }
        router:
          type: string
          nullable: true
          description: |
            Nombre del router. **Editable: no sirve como llave.** Se mantiene
            por compatibilidad; para identificar el grupo use `router_id`.
        router_id:
          type: integer
          nullable: true
          description: Identificador estable del router/grupo lógico.
        router_managed_by_external_aaa:
          type: boolean
          description: |
            `true` = ISPWatch **ya no escribe** en ese equipo. Cubre el
            aprovisionamiento, el corte por mora, la reconexión al pagar y el
            reconciliador: con el interruptor puesto, ISPWatch ordena el cambio
            comercial y publica el evento, pero no toca el router.
        revision: { $ref: "#/components/schemas/Revision" }
        created_at:
          type: string
          nullable: true
          description: |
            Formato `2026-07-03 18:29:20` (UTC, **sin `T` ni `Z`**), a
            diferencia del resto de la API. Ver «Formato de fechas».
          example: "2026-07-03 18:29:20"
        updated_at:
          type: string
          nullable: true
          description: Mismo formato que `created_at`.
          example: "2026-08-18 13:54:59"
    Service:
      type: object
      properties:
        service_id: { type: integer, description: "`user_services.id`. Estable." }
        customer_id: { type: integer }
        status: { $ref: "#/components/schemas/ServiceContractStatus" }
        service_status:
          allOf: [{ $ref: "#/components/schemas/ServiceStatus" }]
          description: |
            Estado comercial del cliente dueño. **Es el que decide si hay
            servicio**, por encima de `status`.
        technology:
          type: string
          enum: [FIBER, RADIO]
        excluded_from_billing: { type: boolean }
        plan: { $ref: "#/components/schemas/Plan" }
        network:
          type: object
          description: |
            Atributos de red del punto. Viven en la ficha del cliente: hoy no
            hay multipunto por abonado.
          properties:
            ip: { type: string, nullable: true }
            pppoe_username: { type: string, nullable: true }
            router_id: { type: integer, nullable: true }
            router_name: { type: string, nullable: true }
            managed_by_external_aaa:
              type: boolean
              description: Igual que `router_managed_by_external_aaa` en `/customers`.
        start_date: { type: string, format: date-time, nullable: true }
        end_date: { type: string, format: date-time, nullable: true }
        installation_date: { type: string, format: date, nullable: true }
        revision: { $ref: "#/components/schemas/Revision" }
        updated_at: { type: string, format: date-time, nullable: true }
...